// src/app/riepilogo/statistiche-stazione-cicerone/statistiche-stazione.interface.ts
/**
 * sezione Be
 */
export interface StatisticheStazioneInterface {
  counter?: number;
  anno?:string;
  giorno?: string;
  tempMin?: string;
  tempMax: string;
  tempMedia: string;
  ventoMax?: string;
  pioggia: string;
  pioggiaMaxEvento?: string;
  umiditaMax?: string;
  umiditaMin?: string;
  pressioneMax?: string;
  pressioneMin?: string;
  pioggiaMese?: string;
  pioggiaAnno?: string;
  tempMinMese?: string;
  tempMaxMese?: string;
  tempMediaMese?: string;
  ventoMaxMese?: string;
  umiditaMaxMese?: string;
}